% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation_value.R
\name{conservation_value}
\alias{conservation_value}
\title{Calculates conservation value metrics}
\usage{
conservation_value(a, b, c)
}
\arguments{
\item{a}{The number of shared species}

\item{b}{The number of species unique to one of the communities}

\item{c}{The number of species unique to the other community}
}
\value{
A list with $m1, $m2 and $m3
}
\description{
This function calculates 3 conservation value metrics:
m1-> the proportion of all species that is shared between two communities
m2-> the proportion of all species that is unique to one of the two communities
m3-> the proportion of the species that occur in one of the communities that is unique to that community
}
\examples{
conservation_value(10,15,3)
 conservation_value(10,15,3)
$m1
[1] 0.3571429

$m2
[1] 0.5357143

$m3
[1] 0.6
}
