% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_coef.R
\name{cv_coef}
\alias{cv_coef}
\title{Extracts conservation value coefficients from a pair a community matr conservation value metricsices}
\usage{
cv_coef(x, y)
}
\arguments{
\item{x}{A community matrix, with species in columns and samples in rows}

\item{y}{A community matrix, with species in columns and samples in rows}
}
\value{
a list with $shared (the number of shared species), $unique_x (the number of species unique to community x) and
$unique_y (the number of species unique to community y)
}
\description{
This function takes two community matrices and extracts 3 coefficients used to calculate conservation value metrics
}
\examples{
data(community_data)

#Select all samples in Evironment 'A'
comm.A<-community_data[which(community_data$environment == "A"),]

# Select only the abundance matrix (columns 1 and 2 contain the 'environment' and 'sample' data)
comm.A <- comm.A[, 3:ncol(comm.A)]

#Select all samples in Evironment 'B'
comm.B<-community_data[which(community_data$environment == "B"),]

# Select only the abundance matrix (columns 1 and 2 contain the 'environment' and 'sample' data)
comm.B <- comm.B[, 3:ncol(comm.B)]

#Calculate coefficients
cv_coef(comm.A, comm.B)
$shared
[1] 29

$unique_x
[1] 0

$unique_y
[1] 1
}
