% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Calculates dissimilarity metrics}
\usage{
dissimilarity(a, b, c)
}
\arguments{
\item{a}{The number of shared species}

\item{b}{The number of species unique to one of the communities}

\item{c}{The number of species unique to the other community}
}
\value{
A list with $sor (the Sorensen index) and $jac (the Jaccard index)
}
\description{
This function calculates dissimilarity metrics between two communities (Sorensen and Jaccard indices).
}
\examples{
dissimilarity(10,15,3)
$sor
[1] 0.4736842

$jac
[1] 0.6428571
}
