% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occasional_abundance.R
\name{occasional_abundance}
\alias{occasional_abundance}
\title{Converts Fahrenheit to Kelvin}
\usage{
occasional_abundance(data, abund_class, dont_apply_to = 0)
}
\arguments{
\item{temp_F}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
\examples{
fahrenheit_to_kelvin(32)
}
