% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occasional_occupancy.R
\name{occasional_occupancy}
\alias{occasional_occupancy}
\title{Converts Fahrenheit to Kelvin}
\usage{
occasional_occupancy(data, occup_class, abund_class = 0,
  dont_apply_to = 0)
}
\arguments{
\item{temp_F}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
\examples{
fahrenheit_to_kelvin(32)
}
