% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_CI.R
\name{plot_graph_CI}
\alias{plot_graph_CI}
\title{Plots a line with the confidence intervals as arrows}
\usage{
plot_graph_CI(x)
}
\arguments{
\item{x}{A matrix with 3 columns. Each row represents one point. The 1st columns contains the point value, while the 2nd and 3rd contain the lower and upper limits of the confidence interval, respectively.}
}
\description{
This function plots a line and the lower and upper limits of the conficende interval for each point.
It can be used to plot the  curves generated by 'conservation_value_seq_CI'.
}
\examples{
data(community_data)

#Select all samples in Evironment 'A'
comm.A<-community_data[which(community_data$environment == "A"),]

# Select only the abundance matrix (columns 1 and 2 contain the 'environment' and 'sample' data)
comm.A <- comm.A[, 3:ncol(comm.A)]

#Select all samples in Evironment 'B'
comm.B<-community_data[which(community_data$environment == "B"),]

# Select only the abundance matrix (columns 1 and 2 contain the 'environment' and 'sample' data)
comm.B <- comm.B[, 3:ncol(comm.B)]

Create 30 bootstrapped communities for each of the two original communities (comm.A and comm.B)
x_boot<-bootstrap(comm.A, 30)
y_boot<-bootstrap(comm.B, 30)

# Sequentially remove abundance classes from both communities and calculate the m3 conservation metric with 95\% confidence intervals

#Plot the m3 metric with confidence intervals
plot_graph_CI(m3)
}
