% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_coef.R
\name{cv_coef}
\alias{cv_coef}
\title{Converts Fahrenheit to Kelvin}
\usage{
cv_coef(x, y)
}
\arguments{
\item{temp_F}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
\examples{
fahrenheit_to_kelvin(32)
}
